+++ python/pycrdt/_base.py
@@ -61,6 +61,7 @@ class BaseDoc:
         *,
         client_id: int | None = None,
         skip_gc: bool | None = None,
+        guid: str | None = None,
         offset_kind: Literal["utf8", "utf16"] | None = None,
         doc: _Doc | None = None,
         Model=None,
@@ -69,7 +70,7 @@ class BaseDoc:
     ) -> None:
         super().__init__(**data)
         if doc is None:
-            doc = _Doc(client_id, skip_gc, offset_kind)
+            doc = _Doc(client_id, skip_gc, guid, offset_kind)
         elif offset_kind is not None and offset_kind != doc.offset_kind:
             raise ValueError(
                 f"offset_kind={offset_kind!r} does not match doc.offset_kind={doc.offset_kind!r}"